ttn_test/ttn_test/requirements.txt
ajax==3.0.0
cachetools==2.1.0
certifi==2018.8.13
chardet==3.0.4
click==6.7
decorator==4.3.0
django-appconf==1.0.2
docutils==0.14
ecdsa==0.13
enum34==1.1.6
Events==0.3
express==1.0
Flask==1.0.2
Flask-MongoKit==0.6
Flask-PyMongo==2.0.1
future==0.16.0
futures==3.2.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
grpcio==1.14.1
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
javascript==2015.32.59
Jinja2==2.10
lockfile==0.12.2
MarkupSafe==1.0
mongokit==0.9.1.1
paho-mqtt==1.3.1
pid==2.2.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pymongo==2.9
python-daemon==2.1.2
python-jose==3.0.0
requests==2.2.1
rsa==3.4.2
service==0.5.1
setproctitle==1.1.10
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
